refactor: extract lifecycle to worker_lifecycle.go
